requirements.txt
Django==1.11.20
django-webpack-loader==0.6.0
djangorestframework==3.9.2
pytz==2019.1
requests==2.21.0
